refactor(BalanceTable): move component to its own directory
